==> STEP1_Grabbing_TMBD_best_20_highest_revenue_data.ipynb <==
# Grabbing Data From TMDB

We grabbed two data sets from TMDB.

1. Best 20 Revenue values films for each year from 1970 to 2019
2. Top rated 1500 films 

To do this, we used API keys.

### Best 20 Revenue Films from 1970 to 2019

# To use api key, import JSON and requests
# Import Pandas to assign values to dataframes
# Import numpy for series

import json
import requests
import pandas as pd
import numpy as np

# Define a function for openning the json
def get_keys(path):
    with open(path) as f:
        return json.load(f)

# Using the function to open and load all keys in this file 
api_keys = get_keys("/Users/ezgi/.secret_the_movie/movie_api.json")

# Setting value of api key as a variable
ds_key_1 = list(api_keys.values())[0]



Now, we begin to grabbing data from website.
First, we create a new list which contains years.

ds_r = requests.get(url) 
# to see the keys and elements of JSON and decide which I need

ds_r.json().keys()
# to see the keys

for key in ds_r.json().keys():
    print(key) # To see the keys and their types
    value = ds_r.json()[key] # To figure out types
    print(type(value)) # To see the type of values
    print('\n\n') # Separate out data

We understood that all other 3 sets of keys are integer, but 'Results' class is list. 
Now, we find which variable is useful for us

my_list = np.arange(1970,2020,1) # my array from 1970 to 2019
my_list

data = pd.DataFrame([])  #create a new blank dataframe to contain all years values

for i in my_list : # loop for changing the years
    url = f'https://api.themoviedb.org/3/discover/movie?api_key={ds_key_1}&primary_release_year={i}&sort_by=revenue.desc'
    ds_r = requests.get(url)
    df = pd.DataFrame.from_dict(ds_r.json()['results'])
    data = pd.DataFrame.append(data,df)

# now, 'data' is our dataframe with values

data.reset_index()     #To assign a new index before converting to csv

data.to_csv('20highest_revenue_from_1970_to_2019')    #write data to csv file and we will use this csv after that

### Top Rated 1500 films

Now, we begin to collect top 1500 films using different URL from TMDB.

data1 = pd.DataFrame([]) #create a new blank dataframe to contain all years values

#assign range to 1 from 75
for i in list(range(1, 75)):  # loop for changing the pages 
    url1 = f'https://api.themoviedb.org/3/movie/top_rated?api_key={ds_key_1}&language=en-US&page={i}'
    ds_r1 = requests.get(url1)
    df1 = pd.DataFrame.from_dict(ds_r1.json()['results'])
    data1 = pd.DataFrame.append(data1,df1)
    
# now, 'data1' is our dataframe with values

data1.reset_index()     #To assign a new index before converting to csv

data.to_csv('all_top_rated_without_revenue.csv') #write data to csv file and we will use this csv after that

When we realized the genres of films are coded by numbers, to find the means of the numbers we used another URL

url_genres = f'https://api.themoviedb.org/3/genre/movie/list?api_key={ds_key_1}&language=en-US' 
# To learn id_code of genres

ds_genres = requests.get(url_genres)

# List of genres

df_genres = pd.DataFrame.from_dict(ds_genres.json()['genres'])
print(len(df_genres)) #Print how many rows
print(df_genres.columns) #Print column names

df_genres 

df_genres.to_csv('genre_codes.csv') # Write our genres codes and id dataframe to csv

After that, we will use our csv files for our project. All of them are available as csv form.
