==> STEP2_getting_budget_with_soup.ipynb <==
# Grabbing Budget&Revenue Values from 'the-numbers.com'

We need budget values for top rates films.
We grabbed them with web scrabing by using Beautiful Soap.

import pandas as pd                #import pandas and numpy for dataframe and series
import numpy as np
import requests
from bs4 import BeautifulSoup      #import request and Beautiful Soap for 

# Grabbing the first page of list

First page of the list URL finishes with /all so we took this value independently.

For other pages we create new for loop according to pattern of webpages URL.

url = "https://www.the-numbers.com/movie/budgets/all"
res = requests.get(url)                               #Assign URL to res for Beautiful Soap

soup = BeautifulSoup(res.text)             

soup                 #to see the results and find the class which we need
#we understood from values the class we need is 'tr'

movies = soup.findAll('tr')          #Assign class values to our movies

dates = []           #create empty list for dates
names = []           #create empty list for names
budget = []          #create empty list for budget
domestic_gross = []  #create empty list for domestic gross
ww_gross = []        #create empty list for worldwide gross

for i in list(range(1, 101)):
    dates.append((movies[i].findAll('td')[1].string))
    names.append((movies[i].findAll('td')[2].string))
    budget.append((movies[i].findAll('td')[3].string))
    domestic_gross.append((movies[i].findAll('td')[4].string))
    ww_gross.append((movies[i].findAll('td')[5].string))
    
#this for loop takes our necessary values from 'td' class and assign new list

df = pd.DataFrame(list(zip(dates, names,budget,domestic_gross,ww_gross)), 
               columns =['Date', 'Name','Budget','Domestic Gross','WW_Gross']) 
df #to see our new dataframe for page 1

#Assign our values to df

# Grabbing other pages values

We figured out that there is a pattern at the end of the URL for each page.

So, we created new list as my_list according to pattern of the URL's end.

my_list = np.arange(101,1901,100)
my_list         # my array according to page URL pattern

data = pd.DataFrame([])          #create a dataframe for my new combined pages

for index in my_list :  
    url_1 = f"https://www.the-numbers.com/movie/budgets/all/{index}"#index is taking from page which ends 101 to 1801
    res = requests.get(url_1)
    soup = BeautifulSoup(res.text)
    movies_1 = soup.findAll('tr')
#create empty lists for new values
    dates_1 = []
    names_1 = []               
    budget_1= []
    domestic_gross_1 = []
    ww_gross_1 = []

    for a_1 in list(range(1, 101)):
        dates_1.append((movies_1[a_1].findAll('td')[1].string))
        names_1.append((movies_1[a_1].findAll('td')[2].string))
        budget_1.append((movies_1[a_1].findAll('td')[3].string))
        domestic_gross_1.append((movies_1[a_1].findAll('td')[4].string))
        ww_gross_1.append((movies_1[a_1].findAll('td')[5].string))


    df_1 = pd.DataFrame(list(zip(dates_1, names_1,budget_1,domestic_gross_1,ww_gross_1)), 
               columns =['Date', 'Name','Budget','Domestic Gross','WW_Gross']) 
    data = pd.DataFrame.append(data,df_1) 


data.tail()  #to see the last values

# Merging the first page and other pages

to_concat = [df, data]
combined_df = pd.concat(to_concat)
#to merge the first URL and others

combined_df.tail()  #to see the tail of the combined dataframe

combined_df.reset_index() #to reset index before converting to csv

combined_df.to_csv('1900movies_budget_domesticgross_WW_gross.csv') #write them to csv after that we use this csv
