==> STEP3_WebScrapping_DataCollectionDocumentation.ipynb <==
# Creating BeautifulSoup objects for HTML

- We grabbed the data from the website www.boxofficemojo.com through webscrapping process.
- We collected the data from years 2015-2019 to analize the data that is useful for our reccomendations
- After collecting the data we have clubbed them and created a single csv file that has all the data pertained to those years

import pandas as pd

import requests
from bs4 import BeautifulSoup

#url = "https://www.boxofficemojo.com/intl/?ref_=bo_nb_ydw_tab"
#W2019="https://www.boxofficemojo.com/year/world/2019/"
#W2018="https://www.boxofficemojo.com/year/world/2018/"
#W2017="https://www.boxofficemojo.com/year/world/2017/"
#W2016="https://www.boxofficemojo.com/year/world/2016/"
W2015="https://www.boxofficemojo.com/year/world/2015/"
res = requests.get(W2015)

#We have taken the data year by year through webscrapping process

print(res.raise_for_status())#raise an error for unsucessfull status
soup = BeautifulSoup(res.text)#beautifulSoup-python package for parsing HTML and XML


print(soup)


res.text #gives raw form of the HTML code

print(soup.prettify())#gives a readble form of the HTML code

# Reading HTML tables for the years 2015-2019

#pd.read_html(url_domestic)[0]
#pd.read_html(W2019)[0]
#pd.read_html(W2018)[0]
# pd.read_html(W2017)[0]
pd.read_html(W2015)[0] 
#Read HTML Tables into list of dataframe objects


# Converting the collected data into CSV files for data merging and cleaning

#df2019 = pd.read_html(url2)[0]
#dfW2019 = pd.read_html(W2019)[0]
# dfW2018 = pd.read_html(W2018)[0]
dfW2015 = pd.read_html(W2015)[0]

#df2019.to_csv('2019_world_wide_box_office.csv')
#dfW2019.to_csv('2019WWBoxOffice.csv')
# dfW2018.to_csv('2018WWBoxOffice.csv')
dfW2015.to_csv('2015WWBoxOffice.csv')
#Converting into CSV files

dfMerge = [dfW2015,dfW2016,dfW2017,dfW2018,dfW2019]
result=pd.concat(dfMerge)#merging the data to convert them into a single CSV file

result.to_csv('2015_2019WorldWideBoxOffice.csv')#Converting the merged data into single CSV file

The above merged CSV file is merged with other data which is obtained through API for a final data set required for the project
