==> STEP4_Merging_4_different_dataframe_and_cleaning.ipynb <==
# Merging 4 different cvs's data

We grabbed 4 different data sets for all dataframe

1. Best 20 Revenue values films for each year            (20highest_revenue_from_1970_to_2019)
2. Top rated 1500 films                                                    (all_top_rated_without_revenue.csv)       

These first two are from TMDB via API key
3. Budget & Worldwide Gross and Domestic                             (1900movies_budget_domesticgross_WW_gross.csv)

Third one is from the-numbers.com via Web Scrabing
4. Foreign and Domestic Revenues from Box Office Mojo                       (2015_2019WorldWideBoxOffice.csv')                              
Forth one is from Box Office via Web Scrabing

import pandas as pd      #import for our dataframe and series 
import numpy as np

df = pd.read_csv('1900movies_budget_domesticgross_WW_gross.csv') #taking values from  the-numbers.com

df   #to see our data frame

df_1 = pd.read_csv('20highest_revenue_from_1970_to_2019')
#take TMDB datas csv   

df_2 = pd.read_csv('all_top_rated_without_revenue.csv')
#take TMDB other data csv

df_2 #to see the dataframe

## Merging and Cleaning Dublicates for TMDB's two dataframes

to_concat = [df_1, df_2]
combined_df = pd.concat(to_concat)
#the columns are same so we can easily merge them

combined_df.head(20) #to see the head

combined_df.tail()#to see the tail

for col in combined_df.columns:
    print(col)                 #to see the column names for choosing useful columns

combined_df.drop(columns = ['Unnamed: 0','backdrop_path','overview','poster_path','video'], inplace=True)
#Drop the unnecessary columns

combined_df #to see the result

combined_df.set_index('title')  #set the title as index for preparing to other merge 

combined_df.sort_values(by='vote_average', ascending = False)  #To see the best vote_average films

combined_df.info()    #to get info about types and shape

combined_df.drop_duplicates(subset='id',inplace=True) #To figure out and drop dublicated rows

combined_df.info()  #to check the duplicated values if present

## Converting Genre Codes to Genre Names

genres_id = pd.read_csv('genre_codes.csv') #taking our genre codes csv

genres_id #Matching the Genre ID with their respective names

genres_id['id'] = genres_id['id'].astype(str) #convert integers of id to string 


genres_id.info() #to check the conversion

### Making a Dictionary for Genres Dataframe

To convert to genre codes to names, we need for loop.
So, we create dictinary as keys are genres_id , values are genre names.

list_of_id = genres_id['id'].to_list()
# Taking keys of genre dataframe and converting them to list

list_of_genre_names = genres_id['name'].to_list()
# Taking values of genre dataframe and converting them to list

my_dict = zip(list_of_id,list_of_genre_names)
genre_dict=dict(my_dict)
# genre_dict createf for loop

for key,value in genre_dict.items():
    combined_df['genre_ids'] = combined_df['genre_ids'].str.replace(key ,value) 
    
#Converting the numbers to names

combined_df.sort_values(by='release_date',ascending=False).head() #to see the convertion

combined_df.set_index('title',inplace=True)
#Seting a new index as title for preparing other merging

combined_df #to see the new index

## Merging Web Scraping Data for Budgets

df.rename(columns = {'Name' : 'title'}, inplace = True)
#before merging we change name to title for merging according to title

df.set_index('title',inplace=True)
# set the title as index for perfect match

combined_outer = pd.merge(combined_df, df, how = 'outer', left_index = True, right_index = True)

#We merge data according to outer shell because we will grab more data later and clean null values

combined_outer

## Merging Data for Worldwide and Domestic Gross

new_df = pd.read_csv('2015_2019WorldWideBoxOffice.csv') #Taking the 4th data set

new_df.rename(columns = {'Release Group' : 'title'}, inplace = True)
#Rename the columns as title for best merging

new_df.set_index('title',inplace=True)
#Seting index as title for merge according to title

new_df #to see last and new df

combined_last = pd.merge(combined_outer, new_df, how = 'outer', left_index = True, right_index = True)
#Merge all 4 data sets

combined_last #to see the last ready dataframe

combined_last['Title'] = combined_last.index

combined_last.reset_index(inplace=True)

## Dropping Unnecessary Columns and Values

combined_last.drop_duplicates(subset='Title',inplace=True)

#first reset index and create new Title column for dropping dublicates

combined_last.columns #to see the unnnecessary columns

combined_last.drop(columns = ['id','original_title','Unnamed: 0_x','Unnamed: 0_y'], inplace=True)
#drop unnecessary columns

combined_last.isna().sum() #to see null values

## Merging Domestic Columns and Worldwide Columns

In our data there are two columns as domestic from two different web scraping. One is named as 'Domestic' and other one is 'Domestic Gross'. We picked them from two different sources. Now, we merge them in one column. Also, we has 'Worldwide' and 'WW_Gross' columns for total revenue and we need to merge them.

combined_last['Domestic'].fillna(combined_last['Domestic Gross'],inplace=True)

combined_last['Worldwide'].fillna(combined_last['WW_Gross'],inplace=True)

combined_last.isna().sum() #when we merge them we have less null values for these columns

combined_last.drop(columns=['Domestic Gross','WW_Gross'],inplace=True)
#Drop unnecessary columns after merging

combined_last.to_csv('last_all_data.csv')
#This is all our merged data

After here, we will use this data file.
